fix: Reorganize ADT module / small fixes
