--- src/core/adt/derive/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  setoid: require('./setoid')
-}
